refactor events.json
