+++ frontend/schema/events.json
@@ -1,17 +1,17 @@
-{
-    "init": {
+[
+    {
         "name": "init",
         "description": "Scene initialization. Triggered when the scene is loaded.",
         "type": "input",
         "fields": []
     },
-    "render": {
+    {
         "name": "render",
         "description": "Render the scene. Triggered according to the frame's refresh interval, or when explicitly requested.",
         "type": "both",
         "fields": []
     },
-    "keyDown": {        
+    {
         "name": "keyDown",
         "description": "When a key is pressed on a keyboard",
         "type": "input",
@@ -30,7 +30,7 @@
             }
         ]
     },
-    "keyUp": {
+    {
         "name": "keyUp",
         "description": "When a key is released on a keyboard",
         "type": "input",
@@ -49,7 +49,7 @@
             }
         ]
     },
-    "mouseMove": {    
+    {
         "name": "mouseMove",
         "description": "When a mouse moves",
         "type": "input",
@@ -68,7 +68,7 @@
             }
         ]
     },
-    "mouseDown": {    
+    {
         "name": "mouseDown",
         "description": "When a mouse button is pressed",
         "type": "input",
@@ -81,7 +81,7 @@
             }
         ]
     },
-    "mouseUp": {   
+    {
         "name": "mouseUp",
         "description": "When a mouse button is released",
         "type": "input",
@@ -94,19 +94,19 @@
             }
         ]
     },
-    "turnOn": {   
+    {
         "name": "turnOn",
         "description": "Turn the screen on",
         "type": "both",
         "fields": []
     },   
-    "turnOff": {   
+    {
         "name": "turnOff",
         "description": "Turn the screen off",
         "type": "both",
         "fields": []
     },
-    "button": {
+    {
         "name": "button",
         "description": "When a GPIO button is pressed or released",
         "type": "input",
@@ -131,23 +131,23 @@
             }
         ]
     },
-    "setSceneState": {
+    {
         "name": "setSceneState",
         "description": "Set the scene's public state. This is not used for private state changes.",
         "type": "output",
         "fields": []
     },
-    "setCurrentScene": {
+    {
         "name": "setCurrentScene",
         "description": "Set the scene's public state. This is not used for private state changes.",
         "type": "output",
         "fields": [
             {
-                "name": "scene",
+                "name": "sceneId",
                 "label": "Scene",
                 "type": "scene",
                 "required": true
             }
         ]
     }
-}
+]
