+++ frontend/src/scenes/frame/panels/Events/eventsLogic.tsx
@@ -1,9 +1,9 @@
 import { actions, kea, path, reducers, selectors } from 'kea'
 import type { eventsLogicType } from './eventsLogicType'
-import _events from './events.json'
+import _events from '../../../../../schema/events.json'
 import { FrameEvent } from '../../../../types'
 
-const events: Record<string, FrameEvent> = _events as any
+const events: FrameEvent[] = _events as any
 
 export const eventsLogic = kea<eventsLogicType>([
   path(['src', 'scenes', 'frame', 'panels', 'Events', 'eventsLogic']),
@@ -25,8 +25,8 @@ export const eventsLogic = kea<eventsLogicType>([
       (s) => [s.tab],
       (tab): FrameEvent[] => {
         return tab === 'listen'
-          ? Object.values(events).filter((event) => event.type === 'input' || event.type === 'both')
-          : Object.values(events).filter((event) => event.type === 'output' || event.type === 'both')
+          ? events.filter((event) => event.type === 'input' || event.type === 'both')
+          : events.filter((event) => event.type === 'output' || event.type === 'both')
       },
     ],
   }),
